Integrated user-dashboard with Flask
#130
