--- code/front-end/user-dashboard/data.js
@@ -1,1785 +1,0 @@
-const el_nimo_data = {
-    max: 100,
-    data: [
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "78.7"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.08",
-            humidity: "76.9"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.08",
-            humidity: "75.9"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.07",
-            humidity: "76.8"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "75.7"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "75.1"
-        },
-        {
-            latitude: "7",
-            longitude: "-147",
-            humidity: "73.6"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.08",
-            humidity: "77.9"
-        },
-        {
-            latitude: "7",
-            longitude: "-147",
-            humidity: "75"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "77.4"
-        },
-        {
-            latitude: "7",
-            longitude: "-147",
-            humidity: "77"
-        },
-        {
-            latitude: "-2.06",
-            longitude: "-140.1",
-            humidity: "78.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "78.9"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "79.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.8"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "81.1"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "88.6"
-        },
-        {
-            latitude: "-2.06",
-            longitude: "-140.08",
-            humidity: "79.7"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "76.4"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "76.8"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "75.9"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "77.8"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "77.8"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "79.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "81.9"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.08",
-            humidity: "78.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "76.4"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "77.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "74.5"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.07",
-            humidity: "79.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "76.1"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "79.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "81.6"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.07",
-            humidity: "80.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "85.2"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.07",
-            humidity: "81.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.1"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.06",
-            humidity: "78.7"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "75.9"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "78.4"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "75.8"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "82"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "83"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "88.6"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.06",
-            humidity: "81.3"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "81.7"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.07",
-            humidity: "83.6"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "85.4"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.08",
-            humidity: "84.8"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "87.7"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.08",
-            humidity: "81.3"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "84.5"
-        },
-        {
-            latitude: "-2.02",
-            longitude: "-140.08",
-            humidity: "80.9"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "85.7"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "80.3"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "82.2"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.08",
-            humidity: "76"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.08",
-            humidity: "77.2"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "86.2"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "84.2"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.07",
-            humidity: "78.1"
-        },
-        {
-            latitude: "6.99",
-            longitude: "-147.17",
-            humidity: "86.6"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.07",
-            humidity: "79.9"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "82.2"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.07",
-            humidity: "78.1"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "85"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.07",
-            humidity: "75.8"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "85.7"
-        },
-        {
-            latitude: "-2.03",
-            longitude: "-140.07",
-            humidity: "79.1"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.18",
-            humidity: "88.5"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "81.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "85.2"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "82"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "85.8"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "80.1"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "86.1"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "76.7"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "85.3"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.07",
-            humidity: "76.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.15",
-            humidity: "84.8"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "75.8"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.15",
-            humidity: "77.3"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "72.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "81.3"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "77.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "79"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "81.6"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "88.3"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.07",
-            humidity: "82.5"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "85.5"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "83"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.18",
-            humidity: "83.4"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "81.9"
-        },
-        {
-            latitude: "6.95",
-            longitude: "-147.18",
-            humidity: "88.6"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "86"
-        },
-        {
-            latitude: "6.95",
-            longitude: "-147.17",
-            humidity: "85.5"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.1",
-            humidity: "83.1"
-        },
-        {
-            latitude: "6.95",
-            longitude: "-147.17",
-            humidity: "82.6"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "84.2"
-        },
-        {
-            latitude: "6.95",
-            longitude: "-147.17",
-            humidity: "84.8"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "81.8"
-        },
-        {
-            latitude: "6.95",
-            longitude: "-147.17",
-            humidity: "88.8"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "93.6"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "84"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.18",
-            humidity: "90.2"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "82.4"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "83.3"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "81"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "87.8"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "81.6"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "82.4"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "83.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "85.7"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "83.6"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.11",
-            humidity: "82.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "88.2"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "81.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "93.4"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "82.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "85.2"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "84.1"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "83"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "85.6"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "82.9"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "83.2"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "86.2"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "85.4"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "87.2"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "87"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "86.1"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "87.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "77.6"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.1",
-            humidity: "89.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "71.9"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "88.7"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "67.1"
-        },
-        {
-            latitude: "-2.05",
-            longitude: "-140.09",
-            humidity: "89.2"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "72.9"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.1",
-            humidity: "91.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.08",
-            humidity: "92.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.4"
-        },
-        {
-            latitude: "-2.04",
-            longitude: "-140.09",
-            humidity: "88.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "86"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.9"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "81.9"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "80.9"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "83.2"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.16",
-            humidity: "85"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.6"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "85.7"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "88.1"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "83.8"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "86.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.1"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "82.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "85.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "85.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "85.6"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "84.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "88.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "85.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "82.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "84.8"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "89.4"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.16",
-            humidity: "87.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "87.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "89"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.15",
-            humidity: "86.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.15",
-            humidity: "85.6"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "84.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "86.9"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "86.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "83.1"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "83.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "83.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.16",
-            humidity: "83.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "79.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.18",
-            humidity: "80.7"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.18",
-            humidity: "85.6"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.18",
-            humidity: "85.3"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "78.9"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "79.9"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "84.8"
-        },
-        {
-            latitude: "6.98",
-            longitude: "-147.17",
-            humidity: "87.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "89.8"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "85.8"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "88.8"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "83"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "82.4"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "87.5"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "87.3"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "85.4"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "85.6"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "85.2"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "85.8"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "88.3"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "88.9"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.17",
-            humidity: "85.6"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "85.4"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "88.5"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "89.3"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.17",
-            humidity: "87.7"
-        },
-        {
-            latitude: "6.97",
-            longitude: "-147.19",
-            humidity: "88.9"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.18",
-            humidity: "89.5"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.18",
-            humidity: "90.1"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.18",
-            humidity: "88.2"
-        },
-        {
-            latitude: "6.96",
-            longitude: "-147.18",
-            humidity: "88.9"
-        },
-        {
-            latitude: "7",
-            longitude: "-147",
-            humidity: "88"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.93",
-            humidity: "87.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "87.9"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.92",
-            humidity: "86.5"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "86.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "84.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "81.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83.7"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "84.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "85.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "85.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "82.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "81.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "85.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "82.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "89.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "86.5"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "86.5"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "83.7"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "84.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "84"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.92",
-            humidity: "83.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "82"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "88.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "87.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "81"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "81.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "82.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "85.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "86.1"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.93",
-            humidity: "85.6"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.92",
-            humidity: "86.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "84.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "86.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "84.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "82.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "82.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83.1"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "82.5"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "81.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "81.7"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "81"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "81.5"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "84.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "82"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "82.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "84.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "87.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "86.5"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "89.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "87.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "87.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "89.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "89.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "88.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "88.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "85.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "81.5"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "79.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "80.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "80.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "78.7"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "78.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "81.4"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.92",
-            humidity: "80"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "87.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "86.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "86.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "86"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.92",
-            humidity: "83.8"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "84.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "84.8"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "82.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "82.3"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "81.8"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "79.2"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "81.7"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "82.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "82.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "81.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "79.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "76.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.91",
-            humidity: "78.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83.1"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83.1"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "82.7"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "81.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "80.1"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "81.5"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "79.2"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "76.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "80.9"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.92",
-            humidity: "84.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "83.8"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "81"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "82.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "80.3"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.92",
-            humidity: "79.1"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "80.9"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "78.7"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "80.3"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.92",
-            humidity: "79.7"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "79.1"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.92",
-            humidity: "80.2"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.94",
-            humidity: "82.2"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.93",
-            humidity: "80.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "81.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "80.4"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.93",
-            humidity: "82.4"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "79.8"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.93",
-            humidity: "76.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "79.3"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "83.2"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.95",
-            humidity: "84.2"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.94",
-            humidity: "84.9"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.95",
-            humidity: "83.6"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.95",
-            humidity: "83.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.95",
-            humidity: "84.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.95",
-            humidity: "83.5"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.94",
-            humidity: "82.2"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.93",
-            humidity: "83.6"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "88.8"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.93",
-            humidity: "87.1"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.94",
-            humidity: "84"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "84.1"
-        },
-        {
-            latitude: "8.07",
-            longitude: "164.97",
-            humidity: "77.3"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.94",
-            humidity: "81.8"
-        },
-        {
-            latitude: "8.07",
-            longitude: "164.98",
-            humidity: "81.3"
-        },
-        {
-            latitude: "0.02",
-            longitude: "-139.94",
-            humidity: "84.1"
-        },
-        {
-            latitude: "1.96",
-            longitude: "165.01",
-            humidity: "89.1"
-        },
-        {
-            latitude: "8.06",
-            longitude: "164.97",
-            humidity: "78.1"
-        },
-        {
-            latitude: "0",
-            longitude: "-139.93",
-            humidity: "86.9"
-        },
-        {
-            latitude: "1.96",
-            longitude: "165.02",
-            humidity: "78.5"
-        },
-        {
-            latitude: "8.07",
-            longitude: "164.98",
-            humidity: "81.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.95",
-            humidity: "80.6"
-        },
-        {
-            latitude: "1.97",
-            longitude: "165.01",
-            humidity: "74.6"
-        },
-        {
-            latitude: "8.07",
-            longitude: "164.97",
-            humidity: "75.8"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.95",
-            humidity: "77.4"
-        },
-        {
-            latitude: "1.97",
-            longitude: "165",
-            humidity: "78.1"
-        },
-        {
-            latitude: "8.07",
-            longitude: "164.97",
-            humidity: "77.3"
-        },
-        {
-            latitude: "0.01",
-            longitude: "-139.92",
-            humidity: "84.2"
-        },
-        {
-            latitude: "1.97",
-            longitude: "165",
-            humidity: "79.7"
-        }
-    ]
-};
