--- code/front-end/user-dashboard/user-dashboard.css
@@ -1,48 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto+Mono:ital,wght@0,100..700;1,100..700&display=swap');
-body{
-    background-color: #092635;
-    color: #9EC8B9;
-    font-family: "roboto mono", monospace;
-    font-size: 30px;
-}
-#container {
-    margin: 20px;
-    display: flex;
-    flex-direction: row;
-    gap: 10px;
-    width: 100%;
-}
-
-/* Set the size of the map container */
-#map {
-    flex-grow: 2;
-    height: calc(100vh - 40px); /* Full height of the viewport minus margins */
-    min-width: 70%; /* 50% of the viewport width */
-    float: left; /* Float the map container to the left */
-}
-
-#image-container {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    box-sizing: border-box;
-    gap: 20px;
-}
-
-/* Style for the image */
-#image-container img {
-   width: 100%;
-   height: 100%;
-}
-
-input{
-    padding: 10px;
-    margin-bottom: 10px;
-    width: 100%;
-    background-color: #9EC8B9;
-    color: #092635;
-    border-radius: 5px;
-    font-family: "Roboto Mono", monospace;
-    font-size: 16px;
-    font-weight: 600;
-}
